+++ build/models/order.js
@@ -150,7 +150,7 @@ var orderModel = /** @class */ (function () {
                         return [4 /*yield*/, database_1.default.connect()];
                     case 1:
                         conn = _a.sent();
-                        sql = 'Delete from orders Where id = ($1)';
+                        sql = "Delete from orders Where id = ($1) returning id, o_date, o_total, user_id";
                         return [4 /*yield*/, conn.query(sql, [id])];
                     case 2:
                         result = _a.sent();
